Fix styleCop issues
